+++ updater-kickassanime.js
@@ -11,11 +11,13 @@ function init() {
 }
 
 function getAnime(name) {
+    //Search for animename in chache
     if (animeCache[name]) {
         animeID = animeCache[name];
         console.log("Found ID in cache: " + animeID);
         insertButton();
     }
+    //Else get it from API
     chrome.runtime.sendMessage(
         {
             type: "GET_ANIME",
@@ -51,6 +53,7 @@ function updateStatus(con, user = "") {
 
 function parseURL(url) {
     //https://www2.kickassanime.rs/anime/world-witches-hasshin-shimasu-162140/episode-03-550318
+    //Get name and episode from URL
     let urlPattern = "https?:\\/\\/.*kickassanime.*\\/anime\\/(.*)-\\d+\\/episode-0*(\\d+)-\\d+";
 
     let res = url.match(urlPattern);
@@ -64,6 +67,7 @@ function parseURL(url) {
 }
 
 function recieveAnime(list) {
+    //Create the HTML ELements needed for User Interaction
     let ul = document.createElement("ul");
     
     for (let elem of list.data) {
@@ -75,6 +79,7 @@ function recieveAnime(list) {
         ul.appendChild(li);
     }
 
+    //Main Div for Anime List
     let div = document.createElement("div");
     div.id = "MAL_UPDATER_DIV_1"
     div.style = "position: absolute;left: 50%;top: 50%;background-color: rgb(33, 33, 33);border: 3px solid rgb(231, 219, 163);padding: 1em 1em 1em 0;z-index: 300000;transform: translate(-50%, -50%);";
@@ -89,6 +94,7 @@ function recieveAnime(list) {
     textbox.placeholder = "If not in List, enter MAL ID here";
     textbox.style = "width: -webkit-fill-available;margin-left: 1.5em;";
 
+    //Div with the two Buttons
     let divButton = document.createElement("div");
     divButton.style = "padding-left: 1.5em;padding-top: 5px;";
 
@@ -116,6 +122,7 @@ function recieveAnime(list) {
 }
 
 function clickedLi(li) {
+    //Save Anime to Cache, insert the Finished button and hide the Div
     animeCache[animeName] = li.value;
     animeID = li.value;
     insertButton();