+++ Changelog.md
@@ -1,5 +1,6 @@
 2.0.1: 27.01.2020
     -Refresh Token is now saved in ms
+    -Bug Fixes
 
 2.0.0: 27.01.2020
     -Added 9anime.to support