+++ Changelog.md
@@ -1,6 +1,7 @@
 2.0.1: 27.01.2020
     -Refresh Token is now saved in ms
     -Bug Fixes
+    -Refresh Token updates after 28 Days
 
 2.0.0: 27.01.2020
     -Added 9anime.to support