+++ Changelog.md
@@ -2,6 +2,7 @@
     -Refresh Token is now saved in ms
     -Bug Fixes
     -Refresh Token updates after 28 Days
+    -Changed github page to darkmode
 
 2.0.0: 27.01.2020
     -Added 9anime.to support