+++ Changelog.md
@@ -1,5 +1,6 @@
 2.0.2: 27.01.2020
     -Fixed Auth Storage not working
+    -Added Code Comments
 
 2.0.1: 27.01.2020
     -Refresh Token is now saved in ms